requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-simple-history==2.3.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
drf-nested-routers==0.91
psycopg2-binary==2.7.5
django-extensions==2.1.2
pillow==5.2.0
social-auth-app-django==2.1.0
django-registration==3.0
django-confy==1.0.4
django-imagekit==4.0.2
